third_party/WebKit/LayoutTests/fast/mediastream/getdisplaymedia.html

<!DOCTYPE HTML>
<html>
<head>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>

promise_test(function() {
  assert_idl_attribute(navigator, 'getDisplayMedia');

  return navigator.getDisplayMedia().then(function(s) {
    fail('getDisplayMedia should have failed');
  }).catch(function(e) {
    assert_equals(e.name, 'NotSupportedError');
  });
}, 'getDisplayMedia() call');

</script>
</body>
</html>
